refactor(logging): unify Logger types so --quiet/--verbose apply to services
- Export a shared Logger interface from src/utils/logger.ts
- Rename internal Logger class to LoggerImpl to avoid name conflict
- Replace src/services/logger.ts with a re-export shim pointing to utils
- Update BaseService to default to the singleton logger from utils
- Update auth-manager.ts and db-retry.ts to import from utils/logger.ts
- Remove meta?: any signatures; all methods now use ...args: unknown[]

Fixes #32
